telegram/checkdomains/requirements.txt
dnspython==2.8.0
python-telegram-bot==22.5
PyYAML==6.0.3
